refactor: 💡 refactor Routes and NavHost
