+++ app/src/main/java/com/rickyhu/hushkeyboard/main/MainApp.kt
@@ -8,7 +8,13 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
 import com.rickyhu.hushkeyboard.data.ThemeOption
+import com.rickyhu.hushkeyboard.home.HomeScreen
+import com.rickyhu.hushkeyboard.introduction.IntroductionScreen
+import com.rickyhu.hushkeyboard.settings.SettingsScreen
 import com.rickyhu.hushkeyboard.theme.HushKeyboardTheme
 import org.koin.androidx.compose.koinViewModel
 
@@ -28,7 +34,27 @@ fun MainApp(viewModel: MainViewModel = koinViewModel()) {
             modifier = Modifier.fillMaxSize(),
             color = MaterialTheme.colorScheme.background,
         ) {
-            MainNavHost()
+            val navController = rememberNavController()
+
+            NavHost(
+                navController = navController,
+                startDestination = Route.Home,
+            ) {
+                composable<Route.Home> {
+                    HomeScreen(
+                        navigateToIntroduction = { navController.navigate(route = Route.Introduction) },
+                        navigateToSettings = { navController.navigate(route = Route.Settings) },
+                    )
+                }
+                composable<Route.Settings> {
+                    SettingsScreen()
+                }
+                composable<Route.Introduction> {
+                    IntroductionScreen(
+                        navigateToHome = { navController.popBackStack() },
+                    )
+                }
+            }
         }
     }
 }
