+++ app/src/main/java/com/rickyhu/hushkeyboard/main/Route.kt
@@ -1,0 +1,14 @@
+package com.rickyhu.hushkeyboard.main
+
+import kotlinx.serialization.Serializable
+
+sealed interface Route {
+    @Serializable
+    object Home : Route
+
+    @Serializable
+    object Settings : Route
+
+    @Serializable
+    object Introduction : Route
+}
